// backend/src/modules/auth/constants/auth.enum.ts
export enum AUTH_KEY {
  FORGOT_PWD = 'FORGOT_PWD',
}

export enum ROLE {
  ADMIN_CONFIGURATOR = 'ADMIN_CONFIGURATOR',
  USER_MANAGEMENT = 'USER_MANAGEMENT',
  SUPPER_ADMIN = 'SUPPER_ADMIN',
  ADMIN_ANALYST = 'ADMIN_ANALYST',
}

export enum EAction {
  ALL_ACTION = '*',
  VIEW_ALL = 'VIEW_ALL',
  VIEW_DETAIL = 'VIEW_DETAIL',
  CREATE = 'CREATE',
  EDIT = 'EDIT',
  DELETE = 'DELETE',
  EXPORT = 'EXPORT',
  IMPORT = 'IMPORT',
  START = 'START',
  STOP = 'STOP',
  TRIGGER = 'TRIGGER',
  CLEAR_CACHE = 'CLEAR_CACHE',
  READ_AUDIT = 'READ_AUDIT',
  PURGE = 'PURGE',
  RESTORE = 'RESTORE',
  TEST_CONNECTION = 'TEST_CONNECTION',
  PAUSE = 'PAUSE',
  CANCEL = 'CANCEL',
  RESUME = 'RESUME',
  SEND_EMAIL = 'SEND_EMAIL',
  SUBMIT = 'SUBMIT',
  APPROVE = 'APPROVE',
  REJECT = 'REJECT',
  REVENUE_REPORT = 'REVENUE_REPORT',
  FILE_REPORT = 'FILE_REPORT',
  READ_ALL = 'READ_ALL',
}

export enum EModule {
  SEND_EMAIL = 'SEND_EMAIL',
  SUBSCRIBER_QUOTA = 'SUBSCRIBER_QUOTA',
  REPORT = 'REPORT',
  SCHEDULER = 'SCHEDULER',
  FILE_LOG = 'FILE_LOG',
  QUOTA = 'QUOTA',
  PROFILE_MANAGEMENT = 'PROFILE_MANAGEMENT',
  ROLE = 'ROLE',
  MODULE = 'MODULE',
  PERMISSION = 'PERMISSION',
  ACTION = 'ACTION',
  USER = 'USER',
  AUTH = 'AUTH',
  ALL_MODULE = '*',
  CYCLE = 'CYCLE',
  SERVER = 'SERVER',
  THRESHOLD = 'THRESHOLD',
  EXPORT='EXPORT',
}

export enum AUTH_URL {}